[4.0] Empty Trash button in clients and newsfeed (#34711)
* remove empty trash in empty state

* newsfeed

+++ administrator/components/com_banners/src/View/Clients/HtmlView.php
@@ -162,7 +162,7 @@ class HtmlView extends BaseHtmlView
 			}
 		}
 
-		if ($this->state->get('filter.state') == -2 && $canDo->get('core.delete'))
+		if (!$this->isEmptyState && $this->state->get('filter.state') == -2 && $canDo->get('core.delete'))
 		{
 			$toolbar->delete('clients.delete')
 				->text('JTOOLBAR_EMPTY_TRASH')
